v23proxy: Migrate example mojom files over
This also demonstrates their use with the new ServiceDescriber in
the work of https://codereview.chromium.org/1345263002/

The Makefile has a few more commands, notably start-v23proxy,
start-echo-client and start-fortune-client. These ought to interact
properly with Vanadium services and clients if the proper REMOTE_ENDPOINT
is passed (unfortunately that was not tested). Additionally, in
order to run more than one of these at a time, you'll need to
delete a LOCK file for a Mojo cache (noted in the Makefile).

This CL has also identified that we will want to add a method to
Mojo's bindings.Encoder that helps to WriteRawBytes.

Change-Id: If3a57d612a0e6d5082013378a647495e59a4603d